fixed types in ProductDto and CartController

+++ src/main/java/com/kodilla/ecommercee/controller/CartController.java
@@ -5,6 +5,7 @@ import com.kodilla.ecommercee.domain.dto.ProductDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -19,31 +20,31 @@ public class CartController {
         return cartId;
     }
     @GetMapping("{cartId}")
-    public CartDto fetchProductsInCart(@PathVariable long cartId){
+    public CartDto fetchProductsInCart(@PathVariable Long cartId){
         List<ProductDto> productList = new ArrayList<>();
-        productList.add(new ProductDto(1,
+        productList.add(new ProductDto(1L,
                 "product 1",
                 "Test",
-                10,
+                BigDecimal.valueOf(10),
                 "1"));
-        productList.add(new ProductDto(2,
+        productList.add(new ProductDto(2L,
                 "product 2",
                 "Test",
-                20,
+                BigDecimal.valueOf(20),
                 "1"));
         CartDto cartDto = new CartDto(productList);
         return cartDto;
     }
     @PutMapping("{cartId}/{productId}")
-    public void addProductToCart(@PathVariable long cartId, @PathVariable long productId){
+    public void addProductToCart(@PathVariable Long cartId, @PathVariable Long productId){
         System.out.println("product " + productId + " added to " + cartId);
     }
     @DeleteMapping("{cartId}/{productId}")
-    public void removeProductFromCart(@PathVariable long cartId, @PathVariable long productId){
+    public void removeProductFromCart(@PathVariable Long cartId, @PathVariable Long productId){
         System.out.println("element " + productId + " removed from " + cartId);
     }
     @PostMapping("{cartId}/create_order")
-    public void createOrder(@PathVariable long cartId){
+    public void createOrder(@PathVariable Long cartId){
         System.out.println("Order created for cart " + cartId);
     }
